use int64_t more

+++ sperl_array.h
@@ -6,7 +6,7 @@
 // Array
 struct SPerl_array {
   void** values;
-  size_t length;
+  int64_t length;
   int64_t capacity;
 };
